scripts/requirements.txt
generative-ai-hub-sdk[all]
flask
jsonify
hdbcli
python-dotenv
typing
flask_limiter
dbutils
PyPDF2
requests
pandas
pdfplumber
pypdf
numpy
logger
opencv-python-headless
jpype1
camelot-py
tabula-py
pytesseract
pillow
openpyxl
fuzzywuzzy
celery
Optional
reportlab
markdown
schedule
gunicorn==23.0.0
flask_cors
sap-xssec
fastapi
flask_wtf
rdflib
python-docx
mammoth